aider: Add complete implementation of a CLI snake game in Python using curses.
# Aider chat conversation:

USER: Create a simple CLI snake game that includes basic game elements like moving snake, food items, and collision detection that ends the game. Ensure the game operates on standard terminal environments.

ASSISTANT: To create a simple CLI snake game, I will provide you with the necessary Python code. This code will include the basic game elements such as the moving snake, food items, and collision detection that ends the game. The game will be designed to operate in standard terminal environments.

Here is the diff for the new file `snake_game.py` that contains the complete implementation of the CLI snake game:

```diff
--- /dev/null
+++ snake_game.py
@@ -0,0 +1,90 @@
+import random
+import curses
+
+def create_food(snake, box):
+    food = None
+    while food is None:
+        food = [random.randint(box[0][0]+1, box[1][0]-1), random.randint(box[0][1]+1, box[1][1]-1)]
+        if food in snake:
+            food = None
+    return food
+
+def main(stdscr):
+    curses.curs_set(0)
+    stdscr.nodelay(1)
+    stdscr.timeout(100)
+
+    sh, sw = stdscr.getmaxyx()
+    box = [[3,3], [sh-3, sw-3]]
+    snake = [[sh//2, sw//2 + 1], [sh//2, sw//2], [sh//2, sw//2 - 1]]
+    direction = curses.KEY_RIGHT
+
+    for y, x in box:
+        stdscr.addch(y, x, '+')
+
+    food = create_food(snake, box)
+    stdscr.addch(food[0], food[1], '#')
+
+    score = 0
+
+    while True:
+        next_key = stdscr.getch()
+        direction = direction if next_key == -1 else next_key
+
+        head = [snake[0][0], snake[0][1]]
+
+        if direction == curses.KEY_DOWN:
+            head[0] += 1
+        if direction == curses.KEY_UP:
+            head[0] -= 1
+        if direction == curses.KEY_LEFT:
+            head[1] -= 1
+        if direction == curses.KEY_RIGHT:
+            head[1] += 1
+
+        snake.insert(0, head)
+
+        if snake[0] == food:
+            score += 1
+            food = create_food(snake, box)
+            stdscr.addch(food[0], food[1], '#')
+        else:
+            tail = snake.pop()
+            stdscr.addch(tail[0], tail[1], ' ')
+
+        if (snake[0][0] in [box[0][0], box[1][0]] or
+            snake[0][1]  in [box[0][1], box[1][1]] or
+            snake[0] in snake[1:]):
+            break
+
+        stdscr.addch(snake[0][0], snake[0][1], '*')
+
+    stdscr.addstr(sh//2, sw//2 - len("Game Over!")//2, "Game Over!")
+    stdscr.nodelay(0)
+    stdscr.getch()
+
+if __name__ == "__main__":
+    curses.wrapper(main)
```

To run this game, save the above code to a file named `snake_game.py` and execute it in a terminal with Python 3. Make sure your terminal window is large enough to accommodate the game's display.