aider: Created a simple snake game using the curses library for text-based user interfaces in a terminal.
# Aider chat conversation:

USER: Develop a simple snake game that can be played in the command line interface. The game should allow the snake to move in different directions, grow in length when consuming food, and handle the game over condition when the snake hits the wall or itself.

ASSISTANT: To create a simple snake game that can be played in the command line interface, we will need to create a new Python file that will contain the game logic and user interface code. Let's call this file `snake_game.py`. Here is the diff for the new file:

```diff
--- /dev/null
+++ snake_game/snake_game.py
@@ ... @@
+import random
+import curses
+
+def create_food(snake, box):
+    food = None
+    while food is None:
+        food = [random.randint(box[0][0]+1, box[1][0]-1), random.randint(box[0][1]+1, box[1][1]-1)]
+        if food in snake:
+            food = None
+    return food
+
+def main(stdscr):
+    curses.curs_set(0)
+    stdscr.nodelay(1)
+    stdscr.timeout(100)
+
+    sh, sw = stdscr.getmaxyx()
+    box = [[3,3], [sh-3, sw-3]]
+    textpad.rectangle(stdscr, box[0][0], box[0][1], box[1][0], box[1][1])
+
+    snake = [[sh//2, sw//2 + 1], [sh//2, sw//2], [sh//2, sw//2 - 1]]
+    direction = curses.KEY_RIGHT
+
+    for y, x in snake:
+        stdscr.addstr(y, x, '#')
+
+    food = create_food(snake, box)
+    stdscr.addstr(food[0], food[1], '*')
+
+    score = 0
+    while True:
+        key = stdscr.getch()
+
+        if key in [curses.KEY_RIGHT, curses.KEY_LEFT, curses.KEY_UP, curses.KEY_DOWN]:
+            direction = key
+
+        head = [snake[0][0], snake[0][1]]
+
+        if direction == curses.KEY_RIGHT:
+            head[1] += 1
+        elif direction == curses.KEY_LEFT:
+            head[1] -= 1
+        elif direction == curses.KEY_UP:
+            head[0] -= 1
+        elif direction == curses.KEY_DOWN:
+            head[0] += 1
+
+        snake.insert(0, head)
+
+        if snake[0] == food:
+            score += 1
+            food = create_food(snake, box)
+            stdscr.addstr(food[0], food[1], '*')
+        else:
+            stdscr.addstr(snake[-1][0], snake[-1][1], ' ')
+            snake.pop()
+
+        if (snake[0][0] in [box[0][0], box[1][0]] or
+            snake[0][1]  in [box[0][1], box[1][1]] or
+            snake[0] in snake[1:]):
+            msg = "Game Over!"
+            stdscr.addstr(sh//2, sw//2-len(msg)//2, msg)
+            stdscr.nodelay(0)
+            stdscr.getch()
+            break
+
+        stdscr.addstr(snake[0][0], snake[0][1], '#')
+
+        stdscr.refresh()
+
+if __name__ == "__main__":
+    curses.wrapper(main)
```

This code creates a simple snake game using the `curses` library, which is suitable for creating text-based user interfaces in a terminal. The game initializes with the snake in the center of the screen and handles real-time keyboard input to change the direction of the snake. Food is generated randomly, and the snake grows when it eats the food. The game ends when the snake hits the wall or itself, displaying a "Game Over!" message.